Added unit tests for trying to access the model system session api with an unauthorized user. Reorganized unit tests from TestModelSystem into TestNode and TestLink within model system editing.
