+++ tests/XTMF2.UnitTests/TestModelSystem.cs
@@ -16,19 +16,15 @@
     You should have received a copy of the GNU General Public License
     along with XTMF2.  If not, see <http://www.gnu.org/licenses/>.
 */
-using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using XTMF2;
-using XTMF2.Editing;
-using XTMF2.Controllers;
+using System;
 using System.Linq;
-using XTMF2.ModelSystemConstruct;
-using static XTMF2.Helper;
 using TestXTMF.Modules;
+using XTMF2.ModelSystemConstruct;
 using XTMF2.RuntimeModules;
 using static TestXTMF.TestHelper;
 
-namespace TestXTMF
+namespace XTMF2
 {
     [TestClass]
     public class TestModelSystem
@@ -282,421 +278,6 @@ namespace TestXTMF
                 Assert.AreEqual(5, ms.GlobalBoundary.Modules.Count);
                 Assert.AreEqual(5, ms.GlobalBoundary.Links.Count);
                 Assert.IsFalse(mSession.AddModelSystemStart(user, ms.GlobalBoundary, "FirstStart", out var start, ref error), error);
-            });
-        }
-
-        [TestMethod]
-        public void UndoAddStart()
-        {
-            RunInModelSystemContext("UndoAddStart", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Starts.Count);
-                Assert.IsTrue(mSession.AddModelSystemStart(user, ms.GlobalBoundary, "Start", out var Start, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Starts.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Starts.Count);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Starts.Count);
-                Assert.AreSame(Start, ms.GlobalBoundary.Starts[0]);
-            });
-        }
-
-        [TestMethod]
-        public void UndoRemoveStart()
-        {
-            RunInModelSystemContext("UndoAddStart", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Starts.Count);
-                Assert.IsTrue(mSession.AddModelSystemStart(user, ms.GlobalBoundary, "Start", out var Start, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Starts.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Starts.Count);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Starts.Count);
-                Assert.AreSame(Start, ms.GlobalBoundary.Starts[0]);
-
-                //now test explicitly removing the start
-                Assert.IsTrue(mSession.RemoveStart(user, Start, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Starts.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Starts.Count);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Starts.Count);
-            });
-        }
-
-        [TestMethod]
-        public void UndoAddNode()
-        {
-            RunInModelSystemContext("UndoAddNode", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Modules.Count);
-                Assert.IsTrue(mSession.AddNode(user, ms.GlobalBoundary, "Start", typeof(BasicParameter<int>),
-                    out var mss, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Modules.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Modules.Count);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Modules.Count);
-                Assert.AreSame(mss, ms.GlobalBoundary.Modules[0]);
-            });
-        }
-
-        [TestMethod]
-        public void UndoRemoveNode()
-        {
-            RunInModelSystemContext("UndoRemoveNode", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Modules.Count);
-                Assert.IsTrue(mSession.AddNode(user, ms.GlobalBoundary, "Start", typeof(BasicParameter<int>),
-                    out var mss, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Modules.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Modules.Count);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Modules.Count);
-                Assert.AreSame(mss, ms.GlobalBoundary.Modules[0]);
-
-                // now remove node explicitly
-                Assert.IsTrue(mSession.RemoveNode(user, mss, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Modules.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Modules.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-            });
-        }
-
-        [TestMethod]
-        public void UndoAddLink()
-        {
-            RunInModelSystemContext("UndoAddLink", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Modules.Count);
-                Assert.IsTrue(mSession.AddNode(user, ms.GlobalBoundary, "Start", typeof(BasicParameter<string>),
-                    out var parameter, ref error), error);
-                Assert.IsTrue(mSession.AddNode(user, ms.GlobalBoundary, "Start", typeof(SimpleParameterModule),
-                    out var module, ref error), error);
-                Assert.AreEqual(2, ms.GlobalBoundary.Modules.Count);
-                Assert.AreEqual(0, ms.GlobalBoundary.Links.Count);
-                Assert.IsTrue(mSession.AddLink(user, module, module.Hooks[0], parameter, out var link, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Links.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Links.Count);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Links.Count);
-                Assert.AreSame(link, ms.GlobalBoundary.Links[0]);
-            });
-        }
-
-        [TestMethod]
-        public void UndoRemoveLink()
-        {
-            RunInModelSystemContext("UndoRemoveLink", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Modules.Count);
-                Assert.IsTrue(mSession.AddNode(user, ms.GlobalBoundary, "Start", typeof(BasicParameter<string>),
-                    out var parameter, ref error), error);
-                Assert.IsTrue(mSession.AddNode(user, ms.GlobalBoundary, "Start", typeof(SimpleParameterModule),
-                    out var module, ref error), error);
-                Assert.AreEqual(2, ms.GlobalBoundary.Modules.Count);
-                Assert.AreEqual(0, ms.GlobalBoundary.Links.Count);
-                Assert.IsTrue(mSession.AddLink(user, module, module.Hooks[0], parameter, out var link, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Links.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Links.Count);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Links.Count);
-                Assert.AreSame(link, ms.GlobalBoundary.Links[0]);
-
-                // now remove the link explicitly
-                Assert.IsTrue(mSession.RemoveLink(user, link, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Links.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Links.Count);
-                Assert.AreSame(link, ms.GlobalBoundary.Links[0]);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Links.Count);
-            });
-        }
-
-        [TestMethod]
-        public void AddSingleLinkToDifferentModule()
-        {
-            /*
-             * This test will try to assign a link between from a single hook to different modules.
-             * This operation should remove the first link and then add the second
-             */
-            RunInModelSystemContext("AddSingleLinkToDifferentModule", (user, pSession, mSession) =>
-            {
-                // initialization
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.IsTrue(mSession.AddModelSystemStart(user, ms.GlobalBoundary, "FirstStart", out var start, ref error), error);
-                Assert.IsTrue(mSession.AddNode(user, ms.GlobalBoundary, "MyMSS", typeof(SimpleTestModule), out var mss1, ref error));
-                Assert.IsTrue(mSession.AddNode(user, ms.GlobalBoundary, "MyMSS", typeof(SimpleTestModule), out var mss2, ref error));
-                Assert.IsTrue(mSession.AddLink(user, start, start.Hooks[0], mss1, out var link1, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Links.Count);
-                // This should not create a new link but move the previous one
-                Assert.IsTrue(mSession.AddLink(user, start, start.Hooks[0], mss2, out var link2, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Links.Count);
-            });
-        }
-
-        [TestMethod]
-        public void AddBoundary()
-        {
-            RunInModelSystemContext("AddBoundary", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Boundaries.Count);
-                Assert.IsTrue(mSession.AddBoundary(user, ms.GlobalBoundary, "UniqueName", out Boundary subB, ref error), error);
-                Assert.IsFalse(mSession.AddBoundary(user, ms.GlobalBoundary, "UniqueName", out Boundary fail1, ref error), "Created a second boundary with the same name!");
-                Assert.AreEqual(1, ms.GlobalBoundary.Boundaries.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Boundaries.Count);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Boundaries.Count);
-                Assert.AreSame(subB, ms.GlobalBoundary.Boundaries[0]);
-                Assert.IsFalse(mSession.AddBoundary(user, ms.GlobalBoundary, "UniqueName", out Boundary fail2, ref error), "Created a second boundary with the same name after redo!");
-            });
-        }
-
-        [TestMethod]
-        public void AddBoundaryNullParent()
-        {
-            RunInModelSystemContext("AddBoundaryNullParent", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Boundaries.Count);
-                Assert.ThrowsException<ArgumentNullException>(() =>
-                {
-                    mSession.AddBoundary(user, null, "UniqueName", out Boundary subB, ref error);
-                });
-            });
-        }
-
-        [TestMethod]
-        public void AddBoundaryNulUser()
-        {
-            RunInModelSystemContext("AddBoundaryNullUser", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Boundaries.Count);
-                Assert.ThrowsException<ArgumentNullException>(() =>
-                {
-                    mSession.AddBoundary(null, ms.GlobalBoundary, "UniqueName", out Boundary subB, ref error);
-                });
-            });
-        }
-
-        [TestMethod]
-        public void RemoveBoundary()
-        {
-            RunInModelSystemContext("RemoveBoundary", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Boundaries.Count);
-                Assert.IsTrue(mSession.AddBoundary(user, ms.GlobalBoundary, "UniqueName", out Boundary subB, ref error), error);
-                Assert.IsFalse(mSession.AddBoundary(user, ms.GlobalBoundary, "UniqueName", out Boundary fail1, ref error), "Created a second boundary with the same name!");
-                Assert.AreEqual(1, ms.GlobalBoundary.Boundaries.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Boundaries.Count);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Boundaries.Count);
-                Assert.AreSame(subB, ms.GlobalBoundary.Boundaries[0]);
-                Assert.IsFalse(mSession.AddBoundary(user, ms.GlobalBoundary, "UniqueName", out Boundary fail2, ref error), "Created a second boundary with the same name after redo!");
-
-                // Now test removing the boundary explicitly
-                Assert.IsTrue(mSession.RemoveBoundary(user, ms.GlobalBoundary, subB, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Boundaries.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Boundaries.Count);
-                Assert.AreSame(subB, ms.GlobalBoundary.Boundaries[0]);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(0, ms.GlobalBoundary.Boundaries.Count);
-            });
-        }
-
-        [TestMethod]
-        public void RemoveBoundaryNullBoundary()
-        {
-            RunInModelSystemContext("RemoveBoundary", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Boundaries.Count);
-                Assert.IsTrue(mSession.AddBoundary(user, ms.GlobalBoundary, "UniqueName", out Boundary subB, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Boundaries.Count);
-
-                // Now test removing the boundary explicitly
-                Assert.ThrowsException<ArgumentNullException>(() =>
-                {
-                    mSession.RemoveBoundary(user, ms.GlobalBoundary, null, ref error);
-                });
-            });
-        }
-
-        [TestMethod]
-        public void RemoveBoundaryNullParent()
-        {
-            RunInModelSystemContext("RemoveBoundary", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Boundaries.Count);
-                Assert.IsTrue(mSession.AddBoundary(user, ms.GlobalBoundary, "UniqueName", out Boundary subB, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Boundaries.Count);
-
-                // Now test removing the boundary explicitly
-                Assert.ThrowsException<ArgumentNullException>(() =>
-                {
-                    mSession.RemoveBoundary(user, null, subB, ref error);
-                });
-            });
-        }
-
-        [TestMethod]
-        public void RemoveBoundaryNullUser()
-        {
-            RunInModelSystemContext("RemoveBoundary", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Boundaries.Count);
-                Assert.IsTrue(mSession.AddBoundary(user, ms.GlobalBoundary, "UniqueName", out Boundary subB, ref error), error);
-                Assert.AreEqual(1, ms.GlobalBoundary.Boundaries.Count);
-
-                // Now test removing the boundary explicitly
-                Assert.ThrowsException<ArgumentNullException>(() =>
-                {
-                    mSession.RemoveBoundary(null, ms.GlobalBoundary, subB, ref error);
-                });
-            });
-        }
-
-        [TestMethod]
-        public void RemoveBoundaryNotInBoundary()
-        {
-            RunInModelSystemContext("RemoveBoundary", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                Assert.AreEqual(0, ms.GlobalBoundary.Boundaries.Count);
-                Assert.IsTrue(mSession.AddBoundary(user, ms.GlobalBoundary, "SubB", out Boundary subB, ref error), error);
-                Assert.IsTrue(mSession.AddBoundary(user, ms.GlobalBoundary, "SubC", out Boundary subC, ref error), error);
-                Assert.IsTrue(mSession.AddBoundary(user, subC, "SubCA", out Boundary subCA, ref error), error);
-                Assert.AreEqual(2, ms.GlobalBoundary.Boundaries.Count);
-                Assert.AreEqual(1, subC.Boundaries.Count);
-
-                Assert.IsFalse(mSession.RemoveBoundary(user, ms.GlobalBoundary, subCA, ref error), "Successfully removed a boundary from a grandparent isntead of failing!");
-            });
-        }
-
-        [TestMethod]
-        public void RemoveLinkToBoundariesThatWereRemoved()
-        {
-            RunInModelSystemContext("RemoveLinkToBoundariesThatWereRemoved", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                var global = ms.GlobalBoundary;
-                // Setup the delete
-                Assert.IsTrue(mSession.AddBoundary(user, global, "ToRemove", out var toRemove, ref error), error);
-                Assert.IsTrue(mSession.AddModelSystemStart(user, global, "Start", out var start, ref error), error);
-                Assert.IsTrue(mSession.AddNode(user, toRemove, "Tricky", typeof(IgnoreResult<string>),
-                    out var tricky, ref error), error);
-                Assert.IsTrue(mSession.AddLink(user, start, start.Hooks[0], tricky, out var link, ref error), error);
-                Assert.AreEqual(1, global.Starts.Count);
-                Assert.AreEqual(1, global.Links.Count);
-                Assert.AreEqual(1, toRemove.Modules.Count);
-
-                // Now remove the boundary and check to make sure the number of links is cleaned up
-                Assert.IsTrue(mSession.RemoveBoundary(user, global, toRemove, ref error), error);
-                Assert.AreEqual(0, global.Links.Count, "We did not remove the link during the remove boundary!");
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(1, global.Links.Count, "The link was not restored after the undo on the remove boundary!");
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(0, global.Links.Count, "We did not remove the link again doing the redo of the remove boundary!");
-            });
-        }
-
-        [TestMethod]
-        public void RemoveMultiLinkToBoundariesThatWereRemoved()
-        {
-            RunInModelSystemContext("RemoveMultiLinkToBoundariesThatWereRemoved", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                var global = ms.GlobalBoundary;
-                // Setup the delete
-                Assert.IsTrue(mSession.AddBoundary(user, global, "ToRemove", out var toRemove, ref error), error);
-                Assert.IsTrue(mSession.AddModelSystemStart(user, global, "Start", out var start, ref error), error);
-                Assert.IsTrue(mSession.AddNode(user, global, "Execute", typeof(Execute),
-                    out var execute, ref error), error);
-                Assert.IsTrue(mSession.AddNode(user, toRemove, "Tricky", typeof(IgnoreResult<string>),
-                    out var tricky, ref error), error);
-                Assert.IsTrue(mSession.AddLink(user, start, start.Hooks[0], execute, out var link, ref error), error);
-                Assert.IsTrue(mSession.AddLink(user, execute, GetHook(execute.Hooks, "To Execute"), tricky, out var link2, ref error), error);
-                Assert.AreEqual(1, global.Starts.Count);
-                Assert.AreEqual(1, global.Modules.Count);
-                Assert.AreEqual(2, global.Links.Count);
-                Assert.AreEqual(1, toRemove.Modules.Count);
-
-                // Now remove the boundary and check to make sure the number of links is cleaned up
-                Assert.IsTrue(mSession.RemoveBoundary(user, global, toRemove, ref error), error);
-                Assert.AreEqual(0, ((MultiLink)global.Links.First(l => l.Origin == execute)).Destinations.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(1, ((MultiLink)global.Links.First(l => l.Origin == execute)).Destinations.Count);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(0, ((MultiLink)global.Links.First(l => l.Origin == execute)).Destinations.Count);
-            });
-        }
-
-        [TestMethod]
-        public void RemoveSingleDestinationInMultiLink()
-        {
-            RunInModelSystemContext("RemoveMultiLinkToBoundariesThatWereRemoved", (user, pSession, mSession) =>
-            {
-                var ms = mSession.ModelSystem;
-                string error = null;
-                var global = ms.GlobalBoundary;
-                Assert.IsTrue(mSession.AddModelSystemStart(user, global, "Start", out var start, ref error), error);
-                Assert.IsTrue(mSession.AddNode(user, global, "Execute", typeof(Execute),
-                    out var execute, ref error), error);
-                Assert.IsTrue(mSession.AddNode(user, global, "Tricky", typeof(IgnoreResult<string>),
-                    out var ignore, ref error), error);
-
-                Assert.IsTrue(mSession.AddLink(user, start, start.Hooks[0], execute, out var link, ref error), error);
-                // Create 2 links from the execute into the ignore
-                Assert.IsTrue(mSession.AddLink(user, execute, GetHook(execute.Hooks, "To Execute"),
-                    execute, out var linkI1, ref error), error);
-                Assert.IsTrue(mSession.AddLink(user, execute, GetHook(execute.Hooks, "To Execute"),
-                    execute, out var _, ref error), error);
-
-
-                Assert.AreEqual(2, ((MultiLink)linkI1).Destinations.Count);
-                Assert.IsTrue(mSession.RemoveLinkDestination(user, linkI1, 0, ref error), error);
-                Assert.AreEqual(1, ((MultiLink)linkI1).Destinations.Count);
-                Assert.IsTrue(mSession.Undo(user, ref error), error);
-                Assert.AreEqual(2, ((MultiLink)linkI1).Destinations.Count);
-                Assert.IsTrue(mSession.Redo(user, ref error), error);
-                Assert.AreEqual(1, ((MultiLink)linkI1).Destinations.Count);
             });
         }
     }
