+++ tests/XTMF2.UnitTests/TestRun.cs
@@ -16,21 +16,16 @@
     You should have received a copy of the GNU General Public License
     along with XTMF2.  If not, see <http://www.gnu.org/licenses/>.
 */
-using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using XTMF2;
-using XTMF2.Editing;
-using XTMF2.Controllers;
-using System.Linq;
-using XTMF2.ModelSystemConstruct;
-using static XTMF2.Helper;
-using TestXTMF.Modules;
 using System.IO;
 using System.Threading;
+using TestXTMF;
+using TestXTMF.Modules;
+using XTMF2.ModelSystemConstruct;
 using XTMF2.RuntimeModules;
 using static TestXTMF.TestHelper;
 
-namespace TestXTMF
+namespace XTMF2
 {
     [TestClass]
     public class TestRun
@@ -42,7 +37,7 @@ namespace TestXTMF
             {
                 TestHelper.CreateRunClient(true, (runBus) =>
                 {
-                    
+
                 });
             });
         }
