+++ tests/XTMF2.UnitTests/XTMF2.UnitTests.csproj
@@ -1,7 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>    
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp2.2</TargetFramework>    
+    <RootNamespace>XTMF2</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
